feat: añadir gestión de estados de inicialización y reconexión en el analizador de caballos
🔧 Se han implementado nuevas funciones en `state.ts` para gestionar los estados de inicialización, reconexión y autenticación del usuario. En `main.ts`, se ha actualizado la lógica para manejar la autenticación y la reconexión de manera más eficiente, evitando inicializaciones redundantes. Además, se han modificado las funciones de análisis de caballos para adaptarse a estos nuevos estados, mejorando la robustez y la experiencia del usuario al interactuar con el analizador de caballos.
