+++ src/content/ui.ts
@@ -1,9 +1,9 @@
 import { CONFIG, debugLog, getConversionDisplayText, type ConversionKey } from './config';
 import { getConvertedPrice } from './api';
-import { getCurrentConversion, setCurrentConversion } from './state';
+import { getCurrentConversion, setCurrentConversion, setIsReconnecting, setHasBeenAuthenticated } from './state';
 import { formatPrice } from './utils/formatting';
 import { createDropdownOptions, createDropdownButton, setupDropdownToggle, type DropdownCallbacks } from './utils/dropdown';
-import { resetAuthenticationState, restartHorseAnalyzer } from './main';
+import { restartHorseAnalyzer } from './main';
 
 // Removed WeakMap cache - elements recreate frequently in SPA navigation
 
@@ -97,13 +97,14 @@ export function createCurrencyConversionUI(ctx: any) {
           const currentBalanceNum = parseFloat(currentBalance) || 0;
           if (lastBalanceNum > 0 && currentBalanceNum === 0) {
             debugLog('User disconnection detected - balance changed from positive to zero');
-            resetAuthenticationState();
+            setHasBeenAuthenticated(false); // Reset authentication state using centralized flag
           }
           
           // Detect user reconnection: balance changes from 0 to positive
           if (lastBalanceNum === 0 && currentBalanceNum > 0) {
             debugLog('User reconnection detected - balance changed from zero to positive');
-            restartHorseAnalyzer(); // Restart the horse analyzer retry loop
+            setIsReconnecting(true); // Set reconnection flag for coordination
+            // The horse analyzer will check this flag and handle reconnection appropriately
           }
           
           lastBalance = currentBalance;
